% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiAnnotator.R
\name{.checkArgsSetDefaults}
\alias{.checkArgsSetDefaults}
\title{Check args and set defaults.}
\usage{
.checkArgsSetDefaults()
}
\description{
This function checks all the arguments passed to an annotation
function and set default values for later use. Evaluation of this function
happens in the parent function.
}
