% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiAnnotator.R
\name{.mergeAndReturn}
\alias{.mergeAndReturn}
\title{Merge results back to the query object and perform additional post
processing steps.}
\usage{
.mergeAndReturn()
}
\description{
This function merges all the calculation results back to the query object.
Additionally, if any flags were set, the function does the necessary checks
and processing to format the return object as required. Evaluation of this
function happens in the parent function.
}
