% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiAnnotator.R
\name{getUCSCtable}
\alias{getUCSCtable}
\title{Obtain a UCSC annotation table given the table & track name.}
\usage{
getUCSCtable(tableName, trackName, bsession = NULL, freeze = "hg18",
  ...)
}
\arguments{
\item{tableName}{Name of the annotation table as it appears on UCSC browser.}

\item{trackName}{Name of the track annotation table as it appears in on
UCSC browser.}

\item{bsession}{UCSC session object returned by \code{\link{makeUCSCsession}}
or \code{\link{browserSession}}. If left NULL the function will call
\code{\link{makeUCSCsession}} with the provided freeze to initiate a session.}

\item{freeze}{one of following: hg18, mm8, rheM, etc. Default is hg18.}

\item{...}{Arguments to be passed to \code{\link{ucscTableQuery}}.}
}
\value{
a dataframe containing the annotation data.
}
\description{
Obtain a UCSC annotation table given the table & track name.
}
\examples{
\dontrun{
refflat <- getUCSCtable("refFlat", "RefSeq Genes")
## same as above ##
refflat <- getUCSCtable("refFlat", "RefSeq Genes",
bsession=session,freeze="hg18")
}
}
\seealso{
\code{\link{makeUCSCsession}}, \code{\link{getNearestFeature}},
\code{\link{getSitesInFeature}}.
}
