\docType{package}
\name{hiAnnotator}
\alias{hiAnnotator}
\alias{hiAnnotator-package}
\title{Annotating RangedData objects with hiAnnotator.}
\description{
  hiAnnotator contains set of functions which allow users
  to annotate a RangedData object with custom set of
  annotations. The basic philosophy of this package is to
  take two RangedData objects (query & subject) with common
  set of space (i.e. chromosomes) and return associated
  annotation per space and rows from the query matching
  space and rows from the subject (i.e. genes or cpg
  islands).
}

