% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiAnnotator.R
\name{makeUCSCsession}
\alias{makeUCSCsession}
\title{Initiate UCSC genome browser session given the freeze argument.}
\usage{
makeUCSCsession(freeze = "hg19")
}
\arguments{
\item{freeze}{one of following: hg19, mm8, rheM, etc. Default is hg19.}
}
\value{
browser session object compatible with rtracklayer functions.
}
\description{
Initiate UCSC genome browser session given the freeze argument.
}
\examples{
\dontrun{
session <- makeUCSCsession()
genome(session)
session <- makeUCSCsession("mm8")
genome(session)
}
}
\seealso{
\code{\link{getUCSCtable}}, \code{\link{makeGRanges}},
\code{\link{getNearestFeature}}, \code{\link{getSitesInFeature}}
}
