\docType{data}
\name{sites}
\alias{sites}
\title{Sample HIV Integration Sites data}
\format{A data frame with 2408 rows and 5 variables}
\description{
  A sample dataset containing collection of unique HIV
  integration sites in the human genome mapped to UCSC
  freeze hg18.
}
\details{
  \itemize{ \item id. A unique id for each row.  \item
  Position. The genomic coordinate of the integration site.
  \item Chr. The chromosome of the integration site.  \item
  Ort. The orientation or strand of the integration site.
  \item sampleName. Patient from which the gDNA sample was
  taken (artificially generated). }
}
\keyword{datasets}

