% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFeatureCountsBig}
\alias{getFeatureCountsBig}
\title{Get counts of annotation within a defined window around each query
range/position for large annotation objects spanning greater than
1 billion rows.}
\usage{
getFeatureCountsBig(sites.rd, features.rd, colnam = NULL, widths = c(1000,
  10000, 1e+06))
}
\arguments{
\item{sites.rd}{GRanges object to be used as the query.}

\item{features.rd}{GRanges object to be used as the subject or the
annotation table.}

\item{colnam}{column name to be added to sites.rd for the newly calculated
annotation...serves as a prefix to windows sizes!}

\item{widths}{a named/numeric vector of window sizes to be used for casting
a net around each position. Default: \code{c(1000,10000,1000000)}}
}
\value{
a GRanges object with new annotation columns appended at the end of
sites.rd. There will be a column for each width defined in widths parameter.
If widths was a named vector i.e. c("100bp"=100,"1K"=1000),
then the colname parameter will be pasted together with width name else
default name will be generated by the function.
}
\description{
Given a query object and window size(s), the function finds all the rows in
subject which are <= window size/2 distance away. Note that here counting is
done using midpoint of the ranges in query instead of start-stop boundaries.
The counts will differ slightly when compared to
\code{\link{getFeatureCounts}}.
}
\examples{
# Convert a dataframe to GRanges object
data(sites)
alldata.rd <- makeGRanges(sites,soloStart=TRUE)

data(genes)
genes.rd <- makeGRanges(genes)

geneCounts1 <- getFeatureCounts(alldata.rd,genes.rd,"NumOfGene")
\dontrun{
geneCounts2 <- getFeatureCountsBig(alldata.rd,genes.rd,"NumOfGene")
identical(geneCounts1,geneCounts2)
}
}
\seealso{
\code{\link{makeGRanges}}, \code{\link{getNearestFeature}},
\code{\link{getSitesInFeature}}, \code{\link{getFeatureCounts}}.
}

