% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getLowestDists}
\alias{getLowestDists}
\title{Get the lowest biological distance from the 5' or 3' boundaries of query
and subject.}
\usage{
getLowestDists(query = NULL, subject = NULL, res.nrst = NULL,
  side = "either", relativeTo = "subject")
}
\arguments{
\item{query}{GRanges object to be used as the query which holds data for
'queryHits' attribute of res.nrst.}

\item{subject}{GRanges object to be used as the subject which holds data for 'subjectHits' attribute of res.nrst.}

\item{res.nrst}{a dataframe of nearest indices as returned by
\code{\link[IRanges]{nearest}}.}

\item{side}{boundary of subject/annotation to use to calculate the
nearest distance. Options are '5p','3p', or the default 'either'.}

\item{relativeTo}{calculate distance relative to query or subject.
Default is 'subject'. See documentation of  \code{\link{getNearestFeature}}
for more information.}
}
\value{
res.nrst with lowest distances appended at the end.
}
\description{
Given a query and subject with indicies from \code{\link[IRanges]{nearest}},
calculate the shortest biological distance to either boundaries of the query
and subject. This is a helper function utilized in
\code{\link{getNearestFeature}}, \code{\link{get2NearestFeature}}
}
\note{
for cases where a query has multiple nearest neighbors or overlaps
with >1 subjects, the function will choose the subject with the lowest
absolute distance.
}
\examples{
query <- GRanges("A", IRanges(c(1, 5, 12, 20), width=1),
strand=c("-","+","-","+"))
subject <- GRanges("A", IRanges(c(1,5,10,15,21), width=8:4),
strand=c("+", "+", "-", "-","-"))
res <- as.data.frame(nearest(query, subject, select="all",
ignore.strand=TRUE))
res <- getLowestDists(query, subject, res, "either", "query")
}
\seealso{
\code{\link{getNearestFeature}}, \code{\link{get2NearestFeature}}.
}

