\docType{package}
\name{hiAnnotator}
\alias{hiAnnotator}
\alias{hiAnnotator-package}
\title{Annotating RangedData or GRanges objects with hiAnnotator.}
\description{
  hiAnnotator contains set of functions which allow users
  to annotate a RangedData or GRanges object with custom
  set of annotations. The basic philosophy of this package
  is to take two RangedData or GRanges objects (query &
  subject) with common set of space/seqnames (i.e.
  chromosomes) and return associated annotation per
  space/seqname and rows from the query matching
  space/seqnames and rows from the subject (i.e. genes or
  cpg islands).
}

