% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiAnnotator.R
\docType{data}
\name{sites}
\alias{sites}
\title{Sample Retrovirus Integration Sites data}
\format{A data frame with 1303 rows and 5 variables}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/?term=12805549}
}
\description{
A sample dataset containing collection of unique HIV & MLV integration sites
in the human genome mapped to UCSC freeze hg18 from PMID: 12805549.
}
\details{
\itemize{
  \item Sequence. Name of the DNA sequence which was aligned to the host
  genome. This is also a unique ID.
  \item Position. The genomic coordinate of the integration site.
  \item Chr. The chromosome of the integration site.
  \item Ort. The orientation or strand of the integration site.
  \item virus. Name of the virus used for the experiment and a given
  sequencing clone.
}
}
\keyword{datasets}

